Move roomNote endpoints into a folder
