+++ server/addRoomNote/index.ts
@@ -1,6 +1,6 @@
 import { AzureFunction, Context, HttpRequest } from '@azure/functions'
 import { authenticatedAzureWrap } from '../src/azureWrap'
-import addRoomNote from '../src/endpoints/addRoomNote'
+import addRoomNote from '../src/endpoints/roomNote/addRoomNote'
 
 const httpTrigger: AzureFunction = async function (context: Context, req: HttpRequest): Promise<void> {
   await authenticatedAzureWrap(context, req, addRoomNote, { audit: true })
