+++ server/deleteRoomNote/index.ts
@@ -1,6 +1,6 @@
 import { AzureFunction, Context, HttpRequest } from '@azure/functions'
 import { authenticatedAzureWrap } from '../src/azureWrap'
-import deleteRoomNote from '../src/endpoints/deleteRoomNote'
+import deleteRoomNote from '../src/endpoints/roomNote/deleteRoomNote'
 
 const httpTrigger: AzureFunction = async function (context: Context, req: HttpRequest): Promise<void> {
   await authenticatedAzureWrap(context, req, deleteRoomNote, { audit: true })
