+++ server/src/endpoints/roomNote/addRoomNote.ts
@@ -1,6 +1,6 @@
-import DB from '../redis'
-import { User } from '../user'
-import { AuthenticatedEndpointFunction, LogFn } from '../endpoint'
+import DB from '../../redis'
+import { User } from '../../user'
+import { AuthenticatedEndpointFunction, LogFn } from '../../endpoint'
 
 const addRoomNote: AuthenticatedEndpointFunction = async (user: User, inputs: any, log: LogFn) => {
   const message = inputs.message
